combined inspecting and extracting from licht file and shared functions into one file.
